backend/requirements.txt
cryptography==43.0.0
flask==3.0.3
flask-sqlalchemy==3.1.1
sqlalchemy-utils==0.41.2
Flask-Migrate==4.0.7
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
marshmallow-union==0.1.15.post1
redis==5.0.8
requests==2.32.3
psycopg2-binary==2.9.9
pytest==8.3.2
factory_boy==3.3.0
gunicorn[gevent]==22.0.0
openai==1.38.0
lark==1.1.9
pytest-asyncio==0.23.8
jsonschema==4.23.0
celery==5.4.0
loguru==0.7.2
pydantic==2.8.2
pyyaml==6.0.1
PyMySQL==1.1.1
boto3==1.34.153
